kenjutsu -> core: Move/renamed.
